fix webpack import warning
